feat: Enhance animations and responsiveness across components
- Updated player_card.dart and stat_widgets.dart to ensure clamped values are converted to double for animation accuracy.
- Expanded goalsight_ui.dart to include new animation exports for better modularity.
- Refactored animated_reveal.dart to utilize GoalSightReveal for improved animation handling.
- Introduced motion_tokens.dart to standardize animation durations and curves.
- Added ai_section_motion.dart for animated AI section transitions.
- Created animated_pressable.dart for enhanced button interactions with scaling and glow effects.
- Developed animations.dart to consolidate animation exports.
- Implemented glow_pulse.dart for pulsing glow effects on components.
- Established reveal_animations.dart for reveal animations with customizable directions and delays.
- Introduced animated_tactical_widgets.dart for animated tactical components including possession bars and charts.
- Created adaptive_layout.dart to manage responsive layouts and adaptive metrics.
- Added responsive.dart to export adaptive layout utilities.
- Developed page_transitions.dart for custom page transitions in navigation.
- Created transitions.dart to export transition utilities for better organization.

+++ app/goal_sight/lib/shared/components/match_card.dart
@@ -320,7 +320,7 @@ class _IntensityBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final clamped = value.clamp(0, 100);
+    final clamped = value.clamp(0, 100).toInt();
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
